# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Internship Recommendation Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentExperienceLevel =
  (typeof StudentExperienceLevel)[keyof typeof StudentExperienceLevel];

export const StudentExperienceLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Student {
  id: number;
  name: string;
  email: string;
  skills: string[];
  domains: string[];
  experienceLevel: StudentExperienceLevel;
  location: string;
  bio?: string;
  careerGoals?: string;
  googleId?: string;
  avatarUrl?: string;
  subscriptionStatus: string;
  isPremium: number;
  resumeText?: string;
  createdAt: string;
}

export type CreateStudentBodyExperienceLevel =
  (typeof CreateStudentBodyExperienceLevel)[keyof typeof CreateStudentBodyExperienceLevel];

export const CreateStudentBodyExperienceLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface CreateStudentBody {
  name: string;
  email: string;
  skills: string[];
  domains: string[];
  experienceLevel: CreateStudentBodyExperienceLevel;
  location: string;
  bio?: string;
  careerGoals?: string;
}

export type UpdateStudentBodyExperienceLevel =
  (typeof UpdateStudentBodyExperienceLevel)[keyof typeof UpdateStudentBodyExperienceLevel];

export const UpdateStudentBodyExperienceLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface UpdateStudentBody {
  name?: string;
  email?: string;
  skills?: string[];
  domains?: string[];
  experienceLevel?: UpdateStudentBodyExperienceLevel;
  location?: string;
  bio?: string;
  careerGoals?: string;
}

export interface Internship {
  id: number;
  title: string;
  company: string;
  domain: string;
  location: string;
  requiredSkills: string[];
  description: string;
  duration: string;
  stipend: string;
  applicationDeadline: string;
  popularity: number;
  createdAt: string;
}

export interface CreateInternshipBody {
  title: string;
  company: string;
  domain: string;
  location: string;
  requiredSkills: string[];
  description: string;
  duration: string;
  stipend: string;
  applicationDeadline: string;
  popularity?: number;
}

export interface RecommendationResult {
  internship: Internship;
  score: number;
  skillMatchScore: number;
  interestScore: number;
  locationScore: number;
  popularityScore: number;
  matchedSkills: string[];
  missingSkills: string[];
  aiInsight: string;
}

export interface CareerPath {
  title: string;
  description: string;
  requiredSkills: string[];
  timelineMonths: number;
  matchScore: number;
  steps: string[];
}

export type SkillGapPriority =
  (typeof SkillGapPriority)[keyof typeof SkillGapPriority];

export const SkillGapPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface SkillGap {
  skill: string;
  priority: SkillGapPriority;
  resources: string[];
}

export interface SkillGapAnalysis {
  studentId: number;
  currentSkills: string[];
  targetRoles: string[];
  skillGaps: SkillGap[];
  recommendations: string[];
}

export interface ActivityItem {
  type: string;
  description: string;
  timestamp: string;
}

export interface DashboardAnalytics {
  totalStudents: number;
  totalInternships: number;
  totalRecommendationsGenerated: number;
  avgMatchScore: number;
  recentActivity: ActivityItem[];
}

export interface DomainStat {
  domain: string;
  count: number;
}

export interface SkillStat {
  skill: string;
  count: number;
}

export type ListInternshipsParams = {
  domain?: string;
  location?: string;
  search?: string;
};

export type GetRecommendationsParams = {
  limit?: number;
};

export type Logout200 = {
  success?: boolean;
};