# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Internship Recommendation Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /internships:
    get:
      operationId: listInternships
      tags: [internships]
      summary: List internships with optional filters
      parameters:
        - name: domain
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Internship"
    post:
      operationId: createInternship
      tags: [internships]
      summary: Create an internship listing (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInternshipBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Internship"
  /internships/{id}:
    get:
      operationId: getInternship
      tags: [internships]
      summary: Get internship by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Internship"
    put:
      operationId: updateInternship
      tags: [internships]
      summary: Update an internship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInternshipBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Internship"
    delete:
      operationId: deleteInternship
      tags: [internships]
      summary: Delete an internship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Internship:
      type: object
      required: [id, title, company, domain, location, requiredSkills, description, duration, stipend, applicationDeadline, popularity, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        domain:
          type: string
        location:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        description:
          type: string
        duration:
          type: string
        stipend:
          type: string
        applicationDeadline:
          type: string
        popularity:
          type: integer
        createdAt:
          type: string
    CreateInternshipBody:
      type: object
      required: [title, company, domain, location, requiredSkills, description, duration, stipend, applicationDeadline]
      properties:
        title:
          type: string
        company:
          type: string
        domain:
          type: string
        location:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        description:
          type: string
        duration:
          type: string
        stipend:
          type: string
        applicationDeadline:
          type: string
        popularity:
          type: integer
...